backend/requirements.txt
pandas
openpyxl
python-dotenv
flask
flask-cors
pymongo
sentence-transformers
numpy
scikit-learn